Use EObjectImpl rather than MinimalEObjectImpl
MinimalEObjectImpl have a lower memory footprint, but every access to its fields (e.g eAdapters or eClass) are heavier and requires to instantiate new Objects.

Reduce memory footprint for Lily is only interesting if we create smaller/less objects in the main loop. In that case, the basic EObjectImpl looks better.